+++ apps/pie-storybook/stories/pie-input.stories.ts
@@ -1,5 +1,7 @@
 import { html } from 'lit';
 import { ifDefined } from 'lit/directives/if-defined.js';
+import { action } from '@storybook/addon-actions';
+import { useArgs as UseArgs } from '@storybook/preview-api';
 
 /* eslint-disable import/no-duplicates */
 import '@justeattakeaway/pie-input';
@@ -7,12 +9,14 @@ import { types, InputProps } from '@justeattakeaway/pie-input';
 /* eslint-enable import/no-duplicates */
 
 import { type StoryMeta } from '../types';
-import { createStory } from '../utilities';
+import { createStory, type TemplateFunction } from '../utilities';
+import '@justeattakeaway/pie-button';
 
 type InputStoryMeta = StoryMeta<InputProps>;
 
 const defaultArgs: InputProps = {
     type: 'text',
+    value: '',
 };
 
 const inputStoryMeta: InputStoryMeta = {
@@ -27,6 +31,13 @@ const inputStoryMeta: InputStoryMeta = {
                 summary: 'text',
             },
         },
+        value: {
+            description: 'The value of the input.',
+            control: 'text',
+            defaultValue: {
+                summary: 'value',
+            },
+        },
     },
     args: defaultArgs,
     parameters: {
@@ -37,10 +48,70 @@ const inputStoryMeta: InputStoryMeta = {
     },
 };
 
-export default inputStoryMeta;
+const Template = ({ type, value }: InputProps) => {
+    const [, updateArgs] = UseArgs();
 
-const Template = ({ type }: InputProps) => html`
-    <pie-input type=${ifDefined(type)}></pie-input>
+    function onInput (event: InputEvent) {
+        const inputElement = event.target as HTMLInputElement;
+        updateArgs({ value: inputElement?.value });
+
+        action('input')({
+            data: event.data,
+            value: inputElement.value,
+        });
+    }
+
+    return html`
+    <pie-input
+        type="${ifDefined(type)}"
+        .value="${value}"
+        @input="${onInput}"></pie-input>
+    `;
+};
+
+const FormTemplate: TemplateFunction<InputProps> = (props: InputProps) => html`
+<h2>Fake form</h2>
+<form id="testForm">
+        <p>Required fields are followed by <strong><span aria-label="required">*</span></strong>.</p>
+        <section>
+            <h2>Contact information</h2>
+            <p>
+                <label for="name">
+                    <span>Name: </span>
+                    <strong><span aria-label="required">*</span></strong>
+                </label>
+                ${Template({ ...props, type: 'text' })}
+            </p>
+        </section>
+        <section style="display: flex; gap: var(--dt-spacing-a); justify-content: flex-end; flex-wrap: wrap; margin-top: var(--dt-spacing-b);">
+            <pie-button type="reset" variant="secondary">Reset</pie-button>
+            <pie-button type="submit" variant="primary">Submit</pie-button>
+        </section>
+    </form>
+    <p id="formLog" style="display: none; font-size: 2rem; color: var(--dt-color-support-positive);"></p>
+    <script>
+        // Display a success message to the user when they submit the form
+        const form = document.querySelector('#testForm');
+        const formLog = document.querySelector('#formLog');
+        form.addEventListener('submit', (e) => {
+            e.preventDefault();
+
+            formLog.innerHTML = 'Form submitted!';
+            formLog.style.display = 'block';
+
+            // Reset the success message after roughly 8 seconds
+            setTimeout(() => {
+                formLog.innerHTML = '';
+                formLog.style.display = 'none';
+            }, 8000);
+        });
+
+    </script>
 `;
 
+const createInputStoryWithForm = createStory<InputProps>(FormTemplate, defaultArgs);
+
 export const Default = createStory<InputProps>(Template, defaultArgs)();
+export const FormIntegration = createInputStoryWithForm();
+
+export default inputStoryMeta;