+++ packages/components/pie-input/src/defs.ts
@@ -5,4 +5,9 @@ export interface InputProps {
      * The type of HTML input to render.
      */
     type?: typeof types[number];
+
+    /**
+     * The value of the input.
+     */
+    value: string;
 }