+++ packages/components/pie-input/src/index.ts
@@ -1,8 +1,11 @@
 import { LitElement, html, unsafeCSS } from 'lit';
 import { property } from 'lit/decorators.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
+import { live } from 'lit/directives/live.js';
 
-import { validPropertyValues, RtlMixin, defineCustomElement } from '@justeattakeaway/pie-webc-core';
+import {
+    validPropertyValues, RtlMixin, defineCustomElement, FormControlMixin,
+} from '@justeattakeaway/pie-webc-core';
 
 import styles from './input.scss?inline';
 import { types, InputProps } from './defs';
@@ -14,19 +17,33 @@ const componentSelector = 'pie-input';
 
 /**
  * @tagname pie-input
+ * @event {InputEvent} input - when the input value is changed.
  */
-export class PieInput extends RtlMixin(LitElement) implements InputProps {
+export class PieInput extends FormControlMixin(RtlMixin(LitElement)) implements InputProps {
     static shadowRootOptions = { ...LitElement.shadowRootOptions, delegatesFocus: true };
 
     @property({ type: String, reflect: true })
     @validPropertyValues(componentSelector, types, 'text')
     public type: InputProps['type'] = 'text';
 
+    @property({ type: String })
+    public value = '';
+
+    /**
+     * Handles data processing in response to the input event. The native input event is left to bubble up.
+     * @param event - The input event.
+     */
+    private handleInput = (event: InputEvent) => {
+        this.value = (event.target as HTMLInputElement).value;
+    };
+
     render () {
-        const { type } = this;
+        const { type, value } = this;
 
         return html`<input
             type=${ifDefined(type)}
+            .value=${live(value)}
+            @input=${this.handleInput}
             data-test-id="pie-input">`;
     }